Renamed vis/layers.js to vis/layers-factory.js
