+++ src/vis/vis.js
@@ -13,7 +13,7 @@ var CartoDBLayerGroupAnonymousMap = require('../geo/cartodb-layer-group-anonymou
 var ModelUpdater = require('./model-updater');
 var LayersCollection = require('../geo/map/layers');
 var AnalysisPoller = require('../analysis/analysis-poller');
-var Layers = require('./layers');
+var LayersFactory = require('./layers-factory');
 
 var STATE_INIT = 'init'; // vis hasn't been sent to Windshaft
 var STATE_OK = 'ok'; // vis has been sent to Windshaft and everything is ok
@@ -334,7 +334,7 @@ var VisModel = Backbone.Model.extend({
     // few moments (e.g: some viz.json files might be cached, etc.).
     var layersData = this._flattenLayers(vizjson.layers);
     return _.map(layersData, function (layerData) {
-      return Layers.create(layerData.type, layerData, layersOptions);
+      return LayersFactory.create(layerData.type, layerData, layersOptions);
     });
   },
